Restore nutrition cold-load budget
