+++ scripts/dom-sink-policy.json
@@ -2,7 +2,7 @@
   "schemaVersion": 1,
   "scope": "Every non-generated JavaScript module under js/",
   "reviewRule": "Any added or modified HTML-writing sink requires security review and a policy refresh.",
-  "scannedFiles": 668,
+  "scannedFiles": 667,
   "sinkCount": 528,
   "files": {
     "js/backup.js": {
